tests(task-service): move tests out so redfish does not need dev-deps
