+++ redfish/src/task_service/mod.rs
@@ -103,96 +103,3 @@ impl<B: Bmc> Resource for TaskService<B> {
         &self.data.as_ref().base
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::error::Error as StdError;
-    use std::io::Error as IoError;
-    use std::io::ErrorKind;
-    use std::sync::Arc;
-
-    use crate::schema::resource::Health as TaskStatus;
-    use crate::schema::task::TaskState;
-    use crate::ServiceRoot;
-    use nv_redfish_bmc_mock::Bmc;
-    use nv_redfish_bmc_mock::Expect;
-    use serde_json::json;
-
-    const TASK_SERVICE_PATH: &str = "/redfish/v1/TaskService";
-    const TASK_PATH: &str = "/redfish/v1/TaskService/Tasks/42";
-
-    #[tokio::test]
-    async fn task_get_exposes_schema_fields() -> Result<(), Box<dyn StdError>> {
-        let bmc = Arc::new(Bmc::<IoError>::default());
-
-        bmc.expect(Expect::get(
-            "/redfish/v1",
-            json!({
-                "@odata.id": "/redfish/v1",
-                "@odata.type": "#ServiceRoot.v1_13_0.ServiceRoot",
-                "Id": "RootService",
-                "Name": "Root Service",
-                "Tasks": {
-                    "@odata.id": TASK_SERVICE_PATH
-                },
-                "Links": {
-                    "Sessions": {
-                        "@odata.id": "/redfish/v1/SessionService/Sessions"
-                    }
-                }
-            }),
-        ));
-
-        bmc.expect(Expect::get(
-            TASK_SERVICE_PATH,
-            json!({
-                "@odata.id": TASK_SERVICE_PATH,
-                "@odata.type": "#TaskService.v1_1_4.TaskService",
-                "Id": "TaskService",
-                "Name": "Task Service",
-                "Tasks": {
-                    "@odata.id": "/redfish/v1/TaskService/Tasks"
-                }
-            }),
-        ));
-
-        bmc.expect(Expect::get(
-            TASK_PATH,
-            json!({
-                "@odata.id": TASK_PATH,
-                "@odata.type": "#Task.v1_4_3.Task",
-                "Id": "42",
-                "Name": "Task 42",
-                "TaskState": "Running",
-                "TaskStatus": "OK",
-                "PercentComplete": 55,
-                "Messages": [{
-                    "MessageId": "Base.1.0.TaskMessage",
-                    "Message": "Task message."
-                }]
-            }),
-        ));
-
-        let root = ServiceRoot::new(bmc).await?;
-        let task_service = root
-            .task_service()
-            .await?
-            .ok_or_else(|| IoError::new(ErrorKind::NotFound, "expected task service"))?;
-        let task = task_service.task(TASK_PATH).await?;
-
-        assert_eq!(task.task_state, Some(TaskState::Running));
-        assert_eq!(task.task_status, Some(TaskStatus::Ok));
-        assert_eq!(task.percent_complete.flatten(), Some(55));
-
-        let messages = task
-            .messages
-            .iter()
-            .flatten()
-            .filter_map(|message| message.message.as_deref())
-            .collect::<Vec<_>>();
-
-        assert_eq!(messages, vec!["Task message."]);
-
-        Ok(())
-    }
-}
